refactor: Rename AddDatasetDialog to AddDatasetDrawer + move JoinBy to graphql/join
